requirements.txt
mlflow
tensorflow
nltk
pillow
matplotlib
numpy
pandas
tqdm
